+++ tests/metro/js/metro.js
@@ -1,7 +1,7 @@
 /*
  * Metro 4 Components Library v4.4.0  (https://metroui.org.ua)
  * Copyright 2012-2020 Sergey Pimenov
- * Built at 17/09/2020 22:43:42
+ * Built at 18/09/2020 14:13:20
  * Licensed under MIT
  */
 (function (global, undefined) {
@@ -603,7 +603,7 @@ function hasProp(obj, prop){
 
 /* global hasProp */
 
-var m4qVersion = "v1.0.8. Built at 27/08/2020 15:05:50";
+var m4qVersion = "v1.0.8. Built at 18/09/2020 13:46:29";
 
 /* eslint-disable-next-line */
 var matches = Element.prototype.matches
@@ -1541,6 +1541,9 @@ $.fn.extend({
 /* global $, not, camelCase, dashedName, isPlainObject, isEmptyObject, isArrayLike, acceptData, parseUnit, getUnit, isVisible, isHidden, matches, strip, normName, hasProp */
 
 $.extend({
+
+    device: (/android|webos|iphone|ipad|ipod|blackberry|iemobile|opera mini/i.test(navigator.userAgent.toLowerCase())),
+
     uniqueId: function (prefix) {
         var d = new Date().getTime();
         if (not(prefix)) {
@@ -4501,7 +4504,7 @@ $.noConflict = function() {
     var Metro = {
 
         version: "4.4.0",
-        compileTime: "17/09/2020 22:43:42",
+        compileTime: "18/09/2020 14:13:20",
         buildNumber: "@@build",
         isTouchable: isTouch,
         fullScreenEnabled: document.fullscreenEnabled,
@@ -18288,6 +18291,222 @@ $.noConflict = function() {
         }
     };
 }(Metro, m4q));
+
+(function(Metro, $) {
+    'use strict';
+
+    var Utils = Metro.utils;
+    var InputMaskDefaultConfig = {
+        maskPattern: ".",
+        mask: null,
+        maskPlaceholder: "_",
+        maskEditableStart: 0,
+        thresholdInterval: 300,
+        onChar: Metro.noop,
+        onInputMaskCreate: Metro.noop
+    };
+
+    Metro.inputMaskSetup = function (options) {
+        InputMaskDefaultConfig = $.extend({}, InputMaskDefaultConfig, options);
+    };
+
+    if (typeof window["metroInputMaskSetup"] !== undefined) {
+        Metro.inputMaskSetup(window["metroInputMaskSetup"]);
+    }
+
+    Metro.Component('input-mask', {
+        init: function( options, elem ) {
+            if ($.device) {
+                console.warn("The component input-mask can't be initialized, because you run it on a mobile device!");
+                return ;
+            }
+            this._super(elem, options, InputMaskDefaultConfig, {
+                // define instance vars here
+                pattern: null,
+                mask: "",
+                maskArray: [],
+                placeholder: "",
+                length: 0,
+                thresholdTimer: null,
+                id: Utils.elementId("input-mask")
+            });
+            return this;
+        },
+
+        _create: function(){
+            this._createStructure();
+            this._createEvents();
+
+            this._fireEvent('input-mask-create');
+        },
+
+        _createStructure: function(){
+            var o = this.options;
+
+            if (!o.mask) {
+                throw new Error('You must provide a pattern for masked input.')
+            }
+
+            if (typeof o.maskPlaceholder !== 'string' || o.maskPlaceholder.length > 1) {
+                throw new Error('Mask placeholder should be a single character or an empty string.')
+            }
+
+            this.placeholder = o.maskPlaceholder;
+            this.mask = (""+o.mask);
+            this.maskArray = this.mask.split("");
+            this.pattern = new RegExp("^"+o.maskPattern+"+$");
+            this.length = this.mask.length;
+
+            this._showValue();
+        },
+
+        _createEvents: function(){
+            var that = this, element = this.element, o = this.options;
+            var editableStart = o.maskEditableStart;
+            var id = this.id;
+
+            var checkEditablePosition = function(pos){
+                if (pos < editableStart) {
+                    setPosition(editableStart);
+                    return false;
+                }
+                return true;
+            }
+
+            var checkEditableChar = function(pos){
+                return pos < that.mask.length && that.mask.charAt(pos) === that.placeholder;
+            }
+
+            var findNextEditablePosition = function (pos){
+                var i, a = that.maskArray;
+
+                for (i = pos; i <= a.length; i++) {
+                    if (a[i] === that.placeholder) {
+                        return i;
+                    }
+                }
+                return pos;
+            }
+
+            var setPosition = function(pos){
+                that.elem.setSelectionRange(pos, pos);
+            }
+
+            var clearThresholdInterval = function(){
+                clearInterval(that.thresholdTimer);
+                that.thresholdTimer = null;
+            }
+
+            element.on("change", function(){
+                if (this.value === "") {
+                    this.value = that.mask;
+                    setPosition(editableStart);
+                }
+            }, {ns: id});
+
+            element.on("focus click", function(){
+                checkEditablePosition(this.selectionStart);
+                setPosition(findNextEditablePosition(this.selectionStart));
+                element.parent().addClass("invalid");
+            }, {ns: id});
+
+            element.on("blur", function(){
+                element.parent().removeClass("invalid");
+            }, {ns: id});
+
+            element.on("keydown", function(e){
+                var pos = this.selectionStart;
+                var val = this.value;
+                var code = e.code, key = e.key;
+
+                if (code === "ArrowRight" || code === "End") {
+                    return true;
+                } else {
+                    if (pos >= that.length && (["Backspace", "Home", "ArrowLeft", "ArrowUp"].indexOf(code) === -1)) {
+                        // Don't move over mask length
+                        e.preventDefault();
+                    } else if (code === "Home" || code === "ArrowUp") {
+                        // Goto editable start position
+                        e.preventDefault();
+                        setPosition(editableStart);
+                    } else if (code === "ArrowLeft") {
+                        if (pos - 1 < editableStart) {
+                            // Don't move behind a editable start position
+                            e.preventDefault();
+                        }
+                    } else if (code === "Backspace") {
+                        e.preventDefault();
+                        if (pos - 1 >= editableStart) {
+                            if (checkEditableChar(pos - 1)) {
+                                if (this.value.charAt(pos - 1) !== that.placeholder) {
+                                    // Replace char if it is not a mask placeholder
+                                    this.value = val.substr(0, pos - 1) + that.placeholder + val.substr(pos);
+                                }
+                            }
+                            // Move to prev char position
+                            setPosition(pos - 1);
+                        }
+                    } else if (code === "Space") {
+                        e.preventDefault();
+                        setPosition(pos + 1);
+                    } else if (!that.pattern.test(key)) {
+                        e.preventDefault();
+                    } else {
+                        e.preventDefault();
+                        if (checkEditableChar(pos)) {
+                            this.value = val.substr(0, pos) + (o.onChar === Metro.noop ? key : Utils.exec(o.onChar, [key], this)) + val.substr(pos + 1);
+                            setPosition(findNextEditablePosition(pos + 1));
+                        }
+                    }
+                }
+            }, {ns: id});
+
+            element.on("keyup", function(){
+                var el = this;
+
+                clearThresholdInterval();
+
+                that.thresholdTimer = setInterval(function(){
+                    clearThresholdInterval();
+                    setPosition(findNextEditablePosition(el.selectionStart));
+                }, o.thresholdInterval)
+            }, {ns: id});
+        },
+
+        _showValue: function(){
+            var that = this, elem = this.elem;
+            var a = new Array(this.length);
+            var val;
+            if (!elem.value) {
+                elem.value = this.mask;
+            } else {
+                val = elem.value;
+                $.each(this.maskArray, function(i, v){
+                    if (val[i] !== v && !that.pattern.test(val[i])) {
+                        a[i] = that.placeholder;
+                    } else {
+                        a[i] = val[i];
+                    }
+                });
+                this.elem.value = a.join("");
+            }
+        },
+
+        destroy: function(){
+            var element = this.element, id = this.id;
+
+            element.off("change", {ns: id});
+            element.off("focus", {ns: id});
+            element.off("click", {ns: id});
+            element.off("blur", {ns: id});
+            element.off("keydown", {ns: id});
+            element.off("keyup", {ns: id});
+
+            return element;
+        }
+    });
+}(Metro, m4q));
+
 
 (function(Metro, $) {
     'use strict';